NEXUS/pyproject.toml
[project]
name = "nexus"
version = "0.1.0"
description = "Multimodal RAG Investigation Platform for Legal Document Intelligence"
requires-python = ">=3.12"
dependencies = [
    # API
    "fastapi>=0.115",
    "uvicorn[standard]",
    "sse-starlette",
    "python-multipart",

    # LLM & AI
    "anthropic>=0.52",
    "openai>=1.60",
    "instructor>=1.14",
    "langchain-core>=0.3",
    "langgraph>=0.4",

    # Document Processing
    "docling>=2.70",
    # unstructured deferred to M3 (email/RTF parsing) — numba dep incompatible with py3.12

    # Retrieval
    "llama-index-core",

    # Vector DB & Graph
    "qdrant-client>=1.12",
    "neo4j>=5.25",
    "graphiti-core>=0.17",

    # NER
    "gliner>=0.2.24",
    "spacy>=3.8",

    # Infrastructure
    "celery[redis]>=5.5",
    "boto3",
    "sqlalchemy>=2.0",
    "alembic",
    "asyncpg",
    "psycopg2-binary",
    "redis>=5.0",
    "aiofiles",

    # Entity Resolution
    "rapidfuzz",

    # Utilities
    "pydantic>=2.10",
    "pydantic-settings",
    "structlog",
    "tenacity",
    "tiktoken",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "workers"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app", "workers"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]
